requirements.txt
streamlit
apify-client
google-generativeai
python-dotenv